requirements.txt
sockets
----------------------------------------------------------------
git
----------------------------------------------------------------
# linux 
$ sudo apt -y install git
----------------------------------------------------------------
customtkinter
# linux: pip3 install customtkinter
----------------------------------------------------------------
Pillow
# linux: python3 -m pip install --upgrade Pillow
----------------------------------------------------------------
tkintermapview
linux: pip3 install tkintermapview
----------------------------------------------------------------
nmap:
$ apt -y install nmap
----------------------------------------------------------------
net-ifaces:
$ sudo apt -y install python-dev
$ pip install netifaces
----------------------------------------------------------------